trading_journal_app/requirements.txt
ttkthemes>=3.2.0
Pillow>=9.0.0
pandas>=1.5.0
matplotlib>=3.5.0
tkcalendar>=1.6.1
seaborn>=0.12.0 